+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3-pre
-*
+* Change exec to display stdout and stderr in chunks
 
 ### 0.0.2
 * Colourise success and error output